tarefa-ebac

+++ src/App.java
@@ -1,7 +1,7 @@
 public class App {
     public static void main(String[] args) {
         int n1 = 20;
-        short n2 = (short) n1;
+        Double n2 = (double) n1;
         System.out.println(n2);
     }
 }
